Separated the session component
