+++ MultigridProjectorClient/MultigridProjectorClient.csproj
@@ -89,6 +89,7 @@
     <ItemGroup>
         <Content Include="Mod\Data\Scripts\MultigridProjectorClient\EmptyMod.cs" />
         <Compile Include="MultigridProjectorPlugin.cs" />
+        <Compile Include="MultigridProjectorSession.cs" />
         <Compile Include="Patches\MyMechanicalConnectionBlockBase_CreateTopPartAndAttach.cs" />
         <Compile Include="Patches\MyMechanicalConnectionBlockBase_RaiseAttachedEntityChanged.cs" />
         <Compile Include="Patches\MyProjectorBase_Build.cs" />
